Remove SpatialJoinGeoUtils, move its helpers to their actual owners
SpatialJoinGeoUtils was a grab-bag of "shared" free functions, but none
of them were actually shared across the base class hierarchy in a way
that justified a separate module:

- getPoint is used by every direct subclass of SpatialJoinAlgorithmBase
  (the S2-based algorithms and RtreeEntryAlgorithm), so it moves onto
  the base class itself.
- prefilterGeoByBoundingBox is only ever reached through
  LibspatialjoinAlgorithm's WKT parsing (SpatialJoinParser.cpp has no
  other caller), so it becomes a static method there.
- getPolyline has nothing to do with any join algorithm or a
  "materialized view" (the old comment was wrong) - it only builds the
  S2 index for the named result cache, so it moves onto
  SpatialJoinCachedIndex, next to its only caller. Its dedicated test
  moves from SpatialJoinAlgorithmsTest.cpp to SpatialJoinCachedIndexTest.cpp.
