+++ src/engine/SpatialJoinParser.cpp
@@ -6,7 +6,7 @@
 
 #include <range/v3/numeric/accumulate.hpp>
 
-#include "engine/spatialJoinAlgorithms/SpatialJoinGeoUtils.h"
+#include "engine/spatialJoinAlgorithms/LibspatialjoinAlgorithm.h"
 
 namespace ad_utility::detail::parallel_wkt_parser {
 
@@ -52,7 +52,7 @@ void WKTParser::processQueue(size_t t) {
         // If we have a prefilter box, check if we also have a precomputed
         // bounding box for the geometry this `VocabIndex` is referring to.
         if (_usePrefiltering &&
-            ad_utility::detail::spatialjoin::prefilterGeoByBoundingBox(
+            LibspatialjoinAlgorithm::prefilterGeoByBoundingBox(
                 _prefilterLatLngBox, _index, job.valueId.getVocabIndex(),
                 job.boundingBox)) {
           prefilterCounter++;
