+++ src/engine/spatialJoinAlgorithms/S2GeometryAlgorithm.cpp
@@ -16,7 +16,6 @@
 #include <s2/s2point_index.h>
 #include <s2/util/units/length-units.h>
 
-#include "engine/spatialJoinAlgorithms/SpatialJoinGeoUtils.h"
 #include "util/GeoConverters.h"
 
 using namespace geometryConverters;
@@ -39,8 +38,7 @@ Result S2GeometryAlgorithm::run() {
 
   // Populate the index
   for (size_t row = 0; row < indexTable->size(); row++) {
-    auto p = ad_utility::detail::spatialjoin::getPoint(indexTable, row,
-                                                       indexJoinCol);
+    auto p = getPoint(indexTable, row, indexJoinCol);
     if (p.has_value()) {
       s2index.Add(toS2Point(p.value()), row);
     }
@@ -64,8 +62,7 @@ Result S2GeometryAlgorithm::run() {
   auto searchJoinCol = indexOfRight ? leftJoinCol : rightJoinCol;
   // Use the index to lookup the points of the other table
   for (size_t searchRow = 0; searchRow < searchTable->size(); searchRow++) {
-    auto p = ad_utility::detail::spatialjoin::getPoint(searchTable, searchRow,
-                                                       searchJoinCol);
+    auto p = getPoint(searchTable, searchRow, searchJoinCol);
     if (!p.has_value()) {
       continue;
     }
