+++ test/engine/SpatialJoinCachedIndexTest.cpp
@@ -4,6 +4,7 @@
 
 #include <gmock/gmock.h>
 #include <s2/mutable_s2shape_index.h>
+#include <s2/s2polyline.h>
 
 #include "../QueryPlannerTestHelpers.h"
 #include "../util/IndexTestHelpers.h"
@@ -13,6 +14,7 @@
 #include "engine/SpatialJoinCachedIndex.h"
 #include "engine/SpatialJoinConfig.h"
 #include "global/ValueId.h"
+#include "index/vocabulary/VocabularyType.h"
 #include "rdfTypes/Variable.h"
 #include "util/Serializer/ByteBufferSerializer.h"
 
@@ -257,5 +259,39 @@ TEST_P(SpatialJoinCachedIndexSimplificationTest, WithSimplification) {
 INSTANTIATE_TEST_SUITE_P(WithAndWithoutSerialization,
                          SpatialJoinCachedIndexSimplificationTest,
                          ::testing::Bool());
+
+// _____________________________________________________________________________
+TEST(SpatialJoinCachedIndex, GetPolylineGeometryTypeCheck) {
+  // Test that `getPolyline` correctly checks the geometry type of its input
+  // literals.
+  std::string kb =
+      "<s1> <asWKT> \"LINESTRING(7.8428469 47.9995367,7.8423373 "
+      "47.9988434,7.8420709 47.9984901,7.8417183 47.9980174,7.8417069 "
+      "47.9980066,7.8413941 47.9975806,7.8413556 47.9975293,7.8413293 "
+      "47.9974942)\"^^<http://www.opengis.net/ont/geosparql#wktLiteral> .\n"
+      "<s2> <asWKT> \"POLYGON((7.8428469 47.9995367,7.8423373 "
+      "47.9988434,7.8420709 47.9984901,7.8417183 47.9980174,7.8417069 "
+      "47.9980066,7.8413941 47.9975806,7.8413556 47.9975293,7.8413293 "
+      "47.9974942, 7.8428469 47.9995367))\""
+      "^^<http://www.opengis.net/ont/geosparql#wktLiteral> .\n"
+      "<s3> <asWKT> \"POINT(1 2)\""
+      "^^<http://www.opengis.net/ont/geosparql#wktLiteral> .\n";
+
+  auto vocabType =
+      ad_utility::VocabularyType::fromString("on-disk-compressed-geo-split");
+  auto qec = ad_utility::testing::getQec(kb, vocabType);
+  auto scan = buildIndexScan(qec, {"?s", std::string{"<asWKT>"}, "?geo"});
+  auto result = scan->getResult();
+  auto col = scan->getVariableColumn(Variable{"?geo"});
+
+  auto check = [&](size_t row) {
+    return SpatialJoinCachedIndex::getPolyline(result->idTableView(), row, col,
+                                               qec->getIndex());
+  };
+
+  EXPECT_TRUE(check(0).has_value());
+  EXPECT_FALSE(check(1).has_value());
+  EXPECT_FALSE(check(2).has_value());
+}
 
 }  // namespace
